spark_agent_project/requirements.txt
mcp[cli]
langchain
langchain-core
langchain-community
langgraph
python-dotenv
# Providers for flexibility
langchain-openai
langchain-anthropic
langchain-google-genai
numpy
